feat: add comprehensive IPv6 and dual-stack PDU session support to SMF
This commit implements full IPv6 and dual-stack (IPv4v6) PDU session support across the SMF, enhancing the 5G core network's capability to handle modern IP connectivity requirements.

Key changes:

- PFCP session handling: Add dynamic IP version detection for F-TEID and outer header creation/removal, supporting both IPv4 and IPv6 GTP-U tunnels across N3/N9 interfaces with 3GPP TS 29.244 compliance
- UE IP address management: Enhance PDR/FAR rules to handle IPv4-only, IPv6-only, and dual-stack UE addresses with proper IPv6 prefix delegation signaling
- Router Advertisement: Implement RFC 4861 compliant Router Advertisement generation for IPv6 stateless address autoconfiguration (SLAAC)
- IP pool architecture: Refactor UE IP pool to support separate IPv4/IPv6 pools with configurable allocation strategies and comprehensive test coverage
- PDU session type handling: Add proper NAS to NGAP PDU session type mapping (IPv4/IPv6/IPv4v6) and session type tracking throughout datapath
- PCO enhancements: Add IPv6 PCSCF request support framework and improve protocol configuration options handling
- Pool data structures: Change LazyReusePool from int to uint64 for extended address space support
- Test coverage: Add extensive unit tests for IPv6 scenarios including dual-stack allocation, router advertisements, and PFCP session modifications

All changes include "WNC:" prefixed logging for easy operational tracing and debugging.

+++ internal/context/ue_ip_pool_test.go
@@ -55,22 +55,22 @@ func TestUeIPPool_ExcludeRange(t *testing.T) {
 		Cidr: "10.10.0.0/24",
 	})
 
-	require.Equal(t, 0x0a0a0000, ueIPPool.Pool().Min())
-	require.Equal(t, 0x0a0a00FF, ueIPPool.Pool().Max())
-	require.Equal(t, 256, ueIPPool.Pool().Remain())
+	require.Equal(t, uint64(0x0a0a0000), ueIPPool.Pool().Min())
+	require.Equal(t, uint64(0x0a0a00FF), ueIPPool.Pool().Max())
+	require.Equal(t, uint64(256), ueIPPool.Pool().Remain())
 
 	excludeUeIPPool := context.NewUEIPPool(&factory.UEIPPool{
 		Cidr: "10.10.0.0/28",
 	})
 
-	require.Equal(t, 0x0a0a0000, excludeUeIPPool.Pool().Min())
-	require.Equal(t, 0x0a0a000F, excludeUeIPPool.Pool().Max())
+	require.Equal(t, uint64(0x0a0a0000), excludeUeIPPool.Pool().Min())
+	require.Equal(t, uint64(0x0a0a000F), excludeUeIPPool.Pool().Max())
 
-	require.Equal(t, 16, excludeUeIPPool.Pool().Remain())
+	require.Equal(t, uint64(16), excludeUeIPPool.Pool().Remain())
 
 	err := ueIPPool.Exclude(excludeUeIPPool)
 	require.NoError(t, err)
-	require.Equal(t, 240, ueIPPool.Pool().Remain())
+	require.Equal(t, uint64(240), ueIPPool.Pool().Remain())
 
 	for i := 16; i <= 255; i++ {
 		allocate := ueIPPool.Allocate(nil)
